works properly

+++ Appman.LeaveManagement/Repositories/ReportingRepository.cs
@@ -15,16 +15,38 @@ namespace Appman.LeaveManagement.Repositories
             _dbContext = dbContext;
         }
 
-        public void AddApprover(Reporting reporting)
+        /// <summary>
+        ///     Add approver of an employee
+        /// </summary>
+        /// <param name="reporting">An instance of Reporting containing the employee's id and approver's id</param>
+        /// <returns>
+        ///     true - if success
+        ///     false - if already exist
+        /// </returns>
+        public bool AddApprover(Reporting reporting)
         {
+            if (_dbContext.Reportings.Contains(reporting))
+                return false;
             _dbContext.Reportings.Add(reporting);
             _dbContext.SaveChanges();
+            return true;
         }
 
-        public void RemoveApprover(Reporting reporting)
+        /// <summary>
+        ///     Remove approver of an employee
+        /// </summary>
+        /// <param name="reporting">An instance of reporting that wants to delete</param>
+        /// <returns>
+        ///     true - if success
+        ///     false - if no reporting match
+        /// </returns>
+        public bool RemoveApprover(Reporting reporting)
         {
+            if (!_dbContext.Reportings.Contains(reporting))
+                return false;
             _dbContext.Reportings.Remove(reporting);
             _dbContext.SaveChanges();
+            return true;
         }
     }
 }
